fastapi_app/requirements.txt
# Description: Python dependencies for FastAPI application.
# Source: Building Enterprise-Ready Data Platforms v2.4, Appendix A.4 (Conceptual).
#
# This version includes dependencies for the Advanced Track with OpenTelemetry.
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
kafka-python==2.0.2
pydantic==2.5.2
# OpenTelemetry packages
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-sdk-extension-aws
opentelemetry-distro
opentelemetry-instrumentation-logging